Implement a FEM residual efficiency noted by Constantine.
The efficiency is in FormFunctionLocalFEM().

By re-ordering the loop over test functions to be inside the loop
over quadrature points, one can reduce evaluations of the nonlinearity
R(u), i.e. the exp() function in Bratu, by a factor of four.

This efficiency can't be had in NGS; the "k" loop over adjacent elements
to the node is over distinct elements.

The result is that Newton-Krylov-multigrid runs are faster by a factor
of about 20%, but there is almost no impact on FAS-multigrid where all
the work is in NGS.

Note I was earlier undercounting flops in FormFunctionLocalFEM(), and
this has been corrected.  Also I made BratuCtx.expcount a PetscLogDouble
to avoid 32-bit integer overflow issues.

TODO: Implementing NJacobi remains.