Important bug fix: ICD-TELE should make sense now.
I don't know what I was thinking when I wrote the
previous version.  This should make sense.  The only way
that the FASCD V-cycle is different, on the upward side
that is, is because it does FAS-type coarse corrections,
whereas the "basic" ICD-TELE iteration refers to the original
operator f(.).