Fix bug in Example 8.2.
The error is seen in the unregularized full cycle results
in the old plap1d.txt file.  Some such runs converged
because of du=0 steps, so _stol was satisfied.
The regularization is needed, at least for full cycles.
In these results, which are better than before, all runs
use -epsreg 1.0-8 (new default) and -fascd_stol 0.